// src/store/mutation.types.ts
export const enum NavType {
  SET_PATH = 'SET_PATH'
}

export const enum RefuelType {
  SET_RULE = 'SET_RULE',
  SET_EXEMPT = 'SET_EXEMPT',
  SET_OIL = 'SET_OIL',
  SET_SEARCH_HISTORY = 'SET_SEARCH_HISTORY',
  CLEAR_SEARCH_HISTORY = 'CLEAR_SEARCH_HISTORY'
}

export const enum CommonType {
  SET_PARTNER_CODE = 'SET_PARTNER_CODE',
  SET_LONGITUDE = 'SET_LONGITUDE',
  SET_LATITUDE = 'SET_LATITUDE',
  SET_SHOW_MONEY = 'SET_SHOW_MONEY',
  SET_HAS_ALLOWANCE = 'SET_HAS_ALLOWANCE',
  SET_CAN_LOAD_MP = 'SET_CAN_LOAD_MP'
}

export const enum UserType {
  SET_MEMBER_ID = 'SET_MEMBER_ID',
  SET_MOBILE = 'SET_MOBILE',
  SET_ACCOUNT = 'SET_ACCOUNT',
  SET_TOKEN = 'SET_TOKEN',
  CLEAR_AUTH = 'CLEAR_AUTH',
  SET_NEWBIE = 'SET_NEWBIE'
}

export const enum SpreadType {
  SET_AGENT_LEVEL = 'SET_AGENT_LEVEL',
  SET_GRANT_STATUS = 'SET_GRANT_STATUS',
  SET_CURRENT_CAPTAIN_ID = 'SET_CURRENT_CAPTAIN_ID'
}